Refactor: Remove pipeline folder
