+++ packages/file-pipeline/src/pipeline.ts
@@ -1,7 +1,7 @@
 import {Writable} from 'stream'
 import File from 'vinyl'
-import {pipeline, through} from '../streams'
-import {Stage, StageArgs, StageConfig} from '../types'
+import {pipeline, through} from './streams'
+import {Stage, StageArgs, StageConfig} from './types'
 import {agnosticSource} from './helpers/agnostic-source'
 import {createEnrichFiles} from './helpers/enrich-files'
 import {createFileCache} from './helpers/file-cache'
